Serialize access to GPIO, reboot on display errors
I was seeing quite frequent issues while updating the display.
Specifically, I would get the following error

    i2c rp2040 peripheral timeout in disable

when calling the `Display()` method, and from this point on the display
would be frozen (though the code was still running, and even `Display()`
calls were being made).

I am pretty sure the cause of that error was simultaneous access to the
I2C interface from different goroutines: one that was updating the
display, another one that was reading new values from the DHT22 sensor.
Avoiding simultaneous accesses to I2C seems to have solved this issues.

I noticed, though, that display errors would still happen sporadically
-- though much less frequently. In particular, I saw this one:

    I2C timeout during write

In order to be super sure, I am now rebooting the device whenever I see
any display issue. This may be a bit too drastic for many of the
possible errors, but in the end of the day rebooting is not a big deal
for this application, so I prefer to be on the safe side and make sure
the device just keeps working, even if I miss a couple measurements
every now and then.