env-server: better Dockerization
First, got rid of that attempt of caching the building of the
dependencies -- it didn't seem to work properly.

Second, make the server port publicly available from the composition.

Third, and here it isn't really about Docker, add configuration specific
for the production environment. For now, we just make sure that the
server binds to 0.0.0.0 (instead of the default 127.0.0.1) in
production.

(To enable production settings, just define ROCKET_PROFILE=production in
the environment.)